Move JournalProxyEditor to proxys.py
